dj_game/urls.py
"""
URL configuration for dj_game project.

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/5.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.contrib import admin
from django.urls import path, include
from game.views import registration, game_view,UsersViewSet, custom_login  # Импортируем представление
from django.contrib.auth import views as auth_views
from rest_framework import routers


router = routers.DefaultRouter()
router.register('players', UsersViewSet)

urlpatterns = [
    path('admin/', admin.site.urls),
    path('register/', registration, name='register'),
    # Страница входа
    path('login/', custom_login, name='login'),
    # Страница выхода
    path('logout/', auth_views.LogoutView.as_view(), name='logout'),
    path('api/', include(router.urls)),
    path('', game_view, name='game'),
]
