requirements.txt
altgraph==0.17.4
asgiref==3.9.1
autopep8==2.3.2
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
Django==4.2.23
djangorestframework==3.16.1
fonttools==4.58.5
gunicorn==23.0.0
idna==3.10
kiwisolver==1.4.8
lxml==6.0.0
markdown-it-py==3.0.0
matplotlib==3.10.3
matplotlib-venn==1.1.2
mdurl==0.1.2
numpy==2.3.1
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
psycopg2-binary==2.9.10
pycodestyle==2.13.0
Pygments==2.19.2
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
pywin32-ctypes==0.2.3
requests==2.32.3
rich==14.0.0
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.4.0
wheel==0.45.1
whitenoise==6.9.0
